Refactor arch specific functions into separate folders
This should simplify adding new architectures.
