+++ examples/fertos_project/src/main.rs
@@ -63,15 +63,15 @@ fn main() -> ! {
         &sc.power_control,
     );
 
-    fe_rtos::interrupt::int_register(
+    fe_rtos::arch::int_register(
         Exception::SysTick.irqn(),
         fe_rtos::task::sys_tick as *const usize,
     );
-    fe_rtos::interrupt::int_register(
+    fe_rtos::arch::int_register(
         Exception::PendSV.irqn(),
         fe_rtos::task::context_switch as *const usize,
     );
-    fe_rtos::interrupt::int_register(
+    fe_rtos::arch::int_register(
         Exception::SVCall.irqn(),
         fe_rtos::syscall::svc_handler as *const usize,
     );
