+++ examples/qemu_project/src/main.rs
@@ -59,15 +59,15 @@ fn spawn_task(_: &mut usize) {
 fn main() -> ! {
     let mut p = cortex_m::peripheral::Peripherals::take().unwrap();
 
-    fe_rtos::interrupt::int_register(
+    fe_rtos::arch::int_register(
         Exception::SysTick.irqn(),
         fe_rtos::task::sys_tick as *const usize,
     );
-    fe_rtos::interrupt::int_register(
+    fe_rtos::arch::int_register(
         Exception::PendSV.irqn(),
         fe_rtos::task::context_switch as *const usize,
     );
-    fe_rtos::interrupt::int_register(
+    fe_rtos::arch::int_register(
         Exception::SVCall.irqn(),
         fe_rtos::syscall::svc_handler as *const usize,
     );
