+++ fe_rtos/build.rs
@@ -12,9 +12,9 @@ fn main() -> Result<(), Box<dyn Error>> {
     println!("cargo:rustc-link-search={}", out_dir.display());
 
     if arch == "arm" {
-        Build::new().file("./src/arm.s").compile("asm");
+        Build::new().file("./src/arch/arm/arm.s").compile("asm");
         Build::new()
-            .file("./src/arm_syscall.s")
+            .file("./src/arch/arm/arm_syscall.s")
             .compile("syscall_asm");
     } else {
         println!("Invalid architecture: {}", arch);
