+++ fe_rtos/src/syscall.rs
@@ -1,7 +1,7 @@
 extern crate alloc;
 
+use crate::arch;
 use crate::fe_alloc;
-use crate::interrupt;
 use crate::ipc;
 use crate::task;
 use alloc::boxed::Box;
@@ -221,6 +221,6 @@ extern "C" fn sys_get_heap_remaining() -> usize {
 
 #[no_mangle]
 extern "C" fn sys_interrupt_register(irqn: isize, handler: *const usize) -> usize {
-    interrupt::int_register(irqn as i8, handler);
+    arch::int_register(irqn as i8, handler);
     0
 }
